# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.3.0
  description: LifeOps personal finance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /incomes:
    get:
      operationId: getIncomes
      tags: [incomes]
      summary: Get all incomes for user
      parameters:
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of incomes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Income"
    post:
      operationId: createIncome
      tags: [incomes]
      summary: Create an income record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncomeRequest"
      responses:
        "201":
          description: Created income
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Income"
  /incomes/{id}:
    put:
      operationId: updateIncome
      tags: [incomes]
      summary: Update an income record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncomeRequest"
      responses:
        "200":
          description: Updated income
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Income"
    delete:
      operationId: deleteIncome
      tags: [incomes]
      summary: Delete an income record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /loans:
    get:
      operationId: getLoans
      tags: [loans]
      summary: Get all loans for user
      responses:
        "200":
          description: List of loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Loan"
    post:
      operationId: createLoan
      tags: [loans]
      summary: Create a loan record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLoanRequest"
      responses:
        "201":
          description: Created loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}:
    put:
      operationId: updateLoan
      tags: [loans]
      summary: Update a loan record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLoanRequest"
      responses:
        "200":
          description: Updated loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
    delete:
      operationId: deleteLoan
      tags: [loans]
      summary: Delete a loan record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /goals:
    get:
      operationId: getGoals
      tags: [goals]
      summary: Get all goals for user
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalRequest"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    put:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalRequest"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Income:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: string
        amount:
          type: number
        date:
          type: string
        notes:
          type: string
        recurrenceType:
          type: string
        occurrences:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, categoryId, amount, date, recurrenceType, occurrences, createdAt, updatedAt]
    CreateIncomeRequest:
      type: object
      properties:
        categoryId:
          type: integer
        amount:
          type: number
        date:
          type: string
        notes:
          type: string
        recurrenceType:
          type: string
        occurrences:
          type: integer
      required: [categoryId, amount, date]
...
    Loan:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        loanType:
          type: string
        principalAmount:
          type: number
        interestRate:
          type: number
        tenureYears:
          type: integer
        emi:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        outstandingBalance:
          type: number
        principalPaid:
          type: number
        interestPaid:
          type: number
        totalInterest:
          type: number
        monthsCompleted:
          type: integer
        monthsRemaining:
          type: integer
        totalMonths:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, name, loanType, principalAmount, interestRate, tenureYears, emi, startDate, endDate, outstandingBalance, principalPaid, interestPaid, totalInterest, monthsCompleted, monthsRemaining, totalMonths, createdAt, updatedAt]
    CreateLoanRequest:
      type: object
      properties:
        name:
          type: string
        loanType:
          type: string
        principalAmount:
          type: number
        interestRate:
          type: number
        tenureYears:
          type: integer
        startDate:
          type: string
        emi:
          type: number
      required: [name, loanType, principalAmount, interestRate, tenureYears, startDate]
    Goal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        targetDate:
          type: string
        progressPercentage:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, name, targetAmount, currentAmount, targetDate, progressPercentage, createdAt, updatedAt]
    CreateGoalRequest:
      type: object
      properties:
        name:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        targetDate:
          type: string
      required: [name, targetAmount, currentAmount, targetDate]
...